Applied the Splash-3 changes
This includes all the synchronization changes, as well as some fixes
related to Splash-2 and the Modified Splash-2. Also, this removes all
the warnings being printed at the moment.

+++ codes/apps/fmm/defs.C
@@ -31,7 +31,7 @@ RoundReal (real val)
 {
    double shifter;
    double frac;
-   long exp;
+   int exp;
    double shifted_frac;
    double new_frac;
    double temp;
